+++ src/ts/AdUnits/Containers/ViewController.ts
@@ -47,7 +47,13 @@ export class ViewController extends AdUnitContainer {
             views = ['videoplayer', 'webview'];
         }
 
-        const orientation = this.getOrientation(options.supportedOrientations, allowRotation, forceOrientation);
+        let orientation = this.getOrientation(options.supportedOrientations, allowRotation, forceOrientation);
+
+        const forcedOrientation = AdUnitContainer.getForcedOrientation();
+        if (forcedOrientation) {
+            allowRotation = false;
+            orientation = forcedOrientation;
+        }
 
         this._nativeBridge.Notification.addNotificationObserver(ViewController._appWillResignActive, []);
         this._nativeBridge.Notification.addAVNotificationObserver(ViewController._audioSessionInterrupt, ['AVAudioSessionInterruptionTypeKey', 'AVAudioSessionInterruptionOptionKey']);