+++ src/ts/Managers/CampaignManager.ts
@@ -21,6 +21,7 @@ import { Campaign } from 'Models/Campaign';
 import { MediationMetaData } from 'Models/MetaData/MediationMetaData';
 import { FrameworkMetaData } from 'Models/MetaData/FrameworkMetaData';
 import { HttpKafka } from 'Utilities/HttpKafka';
+import { SessionManager } from 'Managers/SessionManager';
 
 export class CampaignManager {
 
@@ -29,7 +30,7 @@ export class CampaignManager {
     }
 
     public static setAuctionBaseUrl(baseUrl: string): void {
-        CampaignManager.AuctionBaseUrl = baseUrl + '/v3/games';
+        CampaignManager.AuctionBaseUrl = baseUrl + '/v4/games';
     }
 
     public static setAbGroup(abGroup: number) {
@@ -49,7 +50,7 @@ export class CampaignManager {
     }
 
     private static CampaignBaseUrl: string = 'https://adserver.unityads.unity3d.com/games';
-    private static AuctionBaseUrl: string = 'https://auction.unityads.unity3d.com/v3/games';
+    private static AuctionBaseUrl: string = 'https://auction.unityads.unity3d.com/v4/games';
     private static AbGroup: number | undefined;
     private static CampaignId: string | undefined;
     private static Country: string | undefined;
@@ -68,17 +69,20 @@ export class CampaignManager {
     private _nativeBridge: NativeBridge;
     private _configuration: Configuration;
     private _assetManager: AssetManager;
+    private _sessionManager: SessionManager;
     private _metaDataManager: MetaDataManager;
     private _request: Request;
     private _clientInfo: ClientInfo;
     private _deviceInfo: DeviceInfo;
     private _vastParser: VastParser;
     private _requesting: boolean;
+    private _previousPlacementId: string | undefined;
 
-    constructor(nativeBridge: NativeBridge, configuration: Configuration, assetManager: AssetManager, request: Request, clientInfo: ClientInfo, deviceInfo: DeviceInfo, vastParser: VastParser, metaDataManager: MetaDataManager) {
+    constructor(nativeBridge: NativeBridge, configuration: Configuration, assetManager: AssetManager, sessionManager: SessionManager, request: Request, clientInfo: ClientInfo, deviceInfo: DeviceInfo, vastParser: VastParser, metaDataManager: MetaDataManager) {
         this._nativeBridge = nativeBridge;
         this._configuration = configuration;
         this._assetManager = assetManager;
+        this._sessionManager = sessionManager;
         this._request = request;
         this._clientInfo = clientInfo;
         this._deviceInfo = deviceInfo;
@@ -107,7 +111,7 @@ export class CampaignManager {
                 retryWithConnectionEvents: true
             });
         }).then(response => {
-            if(this._configuration.isPlacementLevelControl()) {
+            if(this._configuration.isAuction()) {
                 return this.parsePlcCampaigns(response);
             } else {
                 return this.parseCampaign(response);
@@ -116,7 +120,7 @@ export class CampaignManager {
             this._requesting = false;
         }).catch((error) => {
             this._requesting = false;
-            if(this._configuration.isPlacementLevelControl()) {
+            if(this._configuration.isAuction()) {
                 this.onPlcError.trigger(error);
             } else {
                 this.onError.trigger(error);
@@ -124,10 +128,25 @@ export class CampaignManager {
         });
     }
 
+    public setPreviousPlacementId(id: string | undefined) {
+        this._previousPlacementId = id;
+    }
+
+    public getPreviousPlacementId(): string | undefined {
+        return this._previousPlacementId;
+    }
+
     private parseCampaign(response: INativeResponse) {
         const json: any = CampaignManager.CampaignResponse ? JsonParser.parse(CampaignManager.CampaignResponse) : JsonParser.parse(response.response);
         if(json.gamerId) {
             this.storeGamerId(json.gamerId);
+        } else if('campaign' in json || 'vast' in json || 'mraid' in json) {
+            this._nativeBridge.Sdk.logError('Unity Ads server returned a campaign without gamerId, ignoring campaign');
+            const error: DiagnosticError = new DiagnosticError(new Error('Missing gamerId'), {
+                rawAdPlan: json
+            });
+            this.onError.trigger(error);
+            return Promise.resolve();
         }
 
         if('campaign' in json) {
@@ -327,7 +346,7 @@ export class CampaignManager {
     private createRequestUrl(): Promise<string> {
         let url: string;
 
-        if(this._configuration.isPlacementLevelControl()) {
+        if(this._configuration.isAuction()) {
             url = [
                 CampaignManager.AuctionBaseUrl,
                 this._clientInfo.getGameId(),
@@ -409,7 +428,7 @@ export class CampaignManager {
                 networkType: this.getParameter('networkType', networkType, 'number'),
             });
 
-            if(this._configuration.isPlacementLevelControl()) {
+            if(this._configuration.isAuction()) {
                 // todo: it's slightly wasteful to read gamerId from storage and then ignore the value
                 url = Url.addParameters(url, {
                     gamerId: this.getParameter('gamerId', this._configuration.getGamerId(), 'string')
@@ -435,20 +454,27 @@ export class CampaignManager {
             bundleId: this.getParameter('bundleId', this._clientInfo.getApplicationName(), 'string'),
             coppa: this.getParameter('coppa', this._configuration.isCoppaCompliant(), 'boolean'),
             language: this.getParameter('language', this._deviceInfo.getLanguage(), 'string'),
+            gameSessionId: this.getParameter('sessionId', this._sessionManager.getGameSessionId(), 'number'),
             timeZone: this.getParameter('timeZone', this._deviceInfo.getTimeZone(), 'string')
         };
+
+        if (this.getPreviousPlacementId()) {
+            body.previousPlacementId = this.getPreviousPlacementId();
+        }
 
         if(typeof navigator !== 'undefined' && navigator.userAgent) {
             body.webviewUa = this.getParameter('webviewUa', navigator.userAgent, 'string');
         }
 
-        if(this._configuration.isPlacementLevelControl()) {
+        if(this._configuration.isAuction()) {
             const placementRequest: any = {};
 
             const placements = this._configuration.getPlacements();
             for(const placement in placements) {
                 if(placements.hasOwnProperty(placement)) {
-                    placementRequest[placement] = {};
+                    placementRequest[placement] = {
+                        adTypes: placements[placement].getAdTypes()
+                    };
                 }
             }